Fix packing of PE files that have import directory entries with 0 functions
This is a "nondestructive" fix because the empty import directories will still exist in both the packed file and an unpacked output of same. It does this by adding an add(const C dll) overload to PeFile::ImportLinker that does not require a function name or ordinal and only adds an empty import entry for the DLL.

Note: only files produced by a seriously broken linker have such imports. The only mainstream product with such a broken linker is Embarcadero {RAD Studio|Delphi|C++ Builder}